requirements.txt
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
openai==1.13.3
anthropic==0.18.1
python-dotenv==1.0.1
backoff==2.2.1
scipy
statsmodels
plotly
altair
bokeh
dask
pyjanitor
xgboost
lightgbm
catboost
tensorflow
keras
torch
holoviews
geopandas
folium
dash
nltk
spacy
transformers
pyarrow
openpyxl
xlrd
pydantic
great_expectations
tsfresh
prophet
networkx
cvxpy
sympy
joblib
numba
h5py
zarr
openai
anthropic
scikit-learn
